requirements.txt
accelerate==0.29.3
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bitsandbytes==0.42.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.2
filelock==3.13.4
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
peft==0.10.0
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scipy==1.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
transformers==4.40.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
websockets==12.0
